src/components/WeatherCurrent.tsx
import {useNavigation} from '@react-navigation/native';
import React, {useCallback, useState} from 'react';
import {LocationService} from '../services/LocationService';
import {Button} from './Button';
import {StyleSheet} from 'react-native';
import {Colors} from '../constants';

export function WeatherCurrent() {
  const [error, setError] = useState(false);
  const [loading, setLoading] = useState(false);
  const navigation = useNavigation();

  const handleFetchWeather = useCallback(async () => {
    setError(false);
    setLoading(true);
    try {
      const position = await LocationService.getCurrentPosition();
      navigation.navigate('Weather', position);
    } catch (e) {
      setError(true);
    }
    setLoading(false);
  }, [navigation]);

  return (
    <Button
      testID="weather-current"
      label="weather at my position"
      onPress={handleFetchWeather}
      loading={loading}
      style={error && styles.error}
    />
  );
}

const styles = StyleSheet.create({
  error: {
    borderColor: Colors.ERROR,
    borderWidth: 1,
    borderRadius: 10,
  },
});
